Keep depreciation and per diem math in BigDecimal
The double-declining factor and first-year service factor were Float;
day-proration in per diem went through Float on its way to a day count.
All exact now.

+++ app/models/per_diem_entry.rb
@@ -26,7 +26,7 @@ class PerDiemEntry < ApplicationRecord
     return 0 if overlap_days <= 0
 
     span_days = (end_date - start_date).to_i + 1
-    [ (qualifying_days.to_i * overlap_days.to_f / span_days).round, qualifying_days.to_i ].min
+    [ (qualifying_days.to_i * overlap_days.to_d / span_days).round, qualifying_days.to_i ].min
   end
 
   def overlaps_period?(period_start, period_end)
